refactor(mocks): split MSW handlers into separate files by domain
Split the monolithic handlers.ts (544 lines) into focused modules:

- handlers.utils.ts: helper function for extracting OpenAI input text
- handlers.openai.ts: OpenAI API endpoints (Responses & Chat Completions)
- handlers.stackone-rpc.ts: StackOne Actions RPC endpoint
- handlers.stackone-hris.ts: StackOne Unified HRIS endpoints
- handlers.stackone-ai.ts: StackOne AI/Tools and OAS spec endpoints
- handlers.example-api.ts: test API endpoints (example.com, petstore)
- handlers.mcp.ts: MCP Protocol endpoints

The main handlers.ts now imports and re-exports all handlers from
these modules. This improves maintainability by grouping related
handlers together and makes it easier to locate specific mock
implementations.
